Refactoring 50%, Arrumando caminho pro sistema de XP

+++ Commands/Perfil.cs
@@ -40,52 +40,60 @@ namespace GTbot.Commands
 		public async Task Imagem(CommandContext ctx, string link)
 		{
 			IMongoDatabase Local = Program.Database.GetDatabase("local");
+			IMongoCollection<Player> Player = Local.GetCollection<Player>("players");
 			IMongoCollection<ClanMember> Member = Local.GetCollection<ClanMember>("membros");
-			IAsyncCursor<ClanMember> query = await Member.FindAsync(x => x.DiscordID == ctx.Member.Id);
-			List<ClanMember> responce = query.ToList();
+			IAsyncCursor<Player> query = await Player.FindAsync(x => x.DiscordID == ctx.Member.Id);
+			List<Player> responce = query.ToList();
 			DiscordEmbedBuilder Embed = new DiscordEmbedBuilder();
-
-			if(responce.Count == 0)
-			{
-				Embed
-					.WithDescription("Só membros do clã possuem acesso ao perfil.")
-					.WithAuthor("Você não é um membro do clã!")
-					.WithColor(Colors.SagiriPink);
-				await ctx.RespondAsync(embed:Embed);
-			}
+			if(responce.Count==0)
+				Player.InsertOne(new Player{DiscordID = ctx.Member.Id,CurrentXP = 0, CurrentLevel = 0, _id = new ObjectId()});
 			else
 			{
-				ClanMember member = responce[0];
-				IMongoCollection<Profile> Profile = Local.GetCollection<Profile>("profile");
-				IAsyncCursor<Profile> query2 = await Profile.FindAsync(x => x.ClanMemberID == member._id);
-				List<Profile> responce2 =  query2.ToList();
+				Player p = responce[0];
+				IAsyncCursor<ClanMember> query2 = await Member.FindAsync(x => x.PlayerID == p._id);
+				List<ClanMember> responce2 = query2.ToList();
 				if(responce2.Count == 0)
 				{
 					Embed
-						.WithDescription("eu criei.\n\nEdite seu perfil novamente para salvar.")
-						.WithAuthor("Você não tinha um perfil, mas ...")
-						.WithColor(Colors.SagiriGray);
+						.WithDescription("Só membros do clã possuem acesso ao perfil.")
+						.WithAuthor("Você não é um membro do clã!")
+						.WithColor(Colors.SagiriPink);
 					await ctx.RespondAsync(embed:Embed);
 				}
 				else
 				{
-					Profile perfil = responce2[0];
-					string b64image = "";
-					using(WebClient wc = new WebClient())
+					ClanMember member = responce2[0];
+					IMongoCollection<Profile> Profile = Local.GetCollection<Profile>("profile");
+					IAsyncCursor<Profile> query3 = await Profile.FindAsync(x => x.ClanMemberID == member._id);
+					List<Profile> responce3 =  query3.ToList();
+					if(responce3.Count == 0)
 					{
-						wc.DownloadFile(link,"_temp.png");
-						byte[] byteStream = File.ReadAllBytes($"{Directory.GetCurrentDirectory()}/_temp.png");
-						b64image = Convert.ToBase64String(byteStream);
-						File.Delete($"{Directory.GetCurrentDirectory()}/_temp.png");
+						Embed
+							.WithDescription("eu criei.\n\nEdite seu perfil novamente para salvar.")
+							.WithAuthor("Você não tinha um perfil, mas ...")
+							.WithColor(Colors.SagiriGray);
+						await ctx.RespondAsync(embed:Embed);
 					}
-					Profile.UpdateOne(Builders<Profile>.Filter.Eq("ClanMemberID",member._id), Builders<Profile>.Update.Set("CustomImage",b64image));
-					Embed
-						.WithAuthor("Imagem atualizada")
-						.WithColor(Colors.SagiriBlue);
-					await ctx.RespondAsync(embed:Embed);
+					else
+					{
+						Profile perfil = responce3[0];
+						string b64image = "";
+						using(WebClient wc = new WebClient())
+						{
+							wc.DownloadFile(link,"_temp.png");
+							byte[] byteStream = File.ReadAllBytes($"{Directory.GetCurrentDirectory()}/_temp.png");
+							b64image = Convert.ToBase64String(byteStream);
+							File.Delete($"{Directory.GetCurrentDirectory()}/_temp.png");
+						}
+						Profile.UpdateOne(Builders<Profile>.Filter.Eq("ClanMemberID",member._id), Builders<Profile>.Update.Set("CustomImage",b64image));
+						Embed
+							.WithAuthor("Imagem atualizada")
+							.WithColor(Colors.SagiriBlue);
+						await ctx.RespondAsync(embed:Embed);
 
-				}	
-			}
+					}	
+				}
+			}	
 		}
 
 		[Command("cor")]
@@ -94,84 +102,94 @@ namespace GTbot.Commands
 			IMongoDatabase Local = Program.Database.GetDatabase("local");
 			DiscordEmbedBuilder Embed = new DiscordEmbedBuilder();
 			IMongoCollection<ClanMember> Member = Local.GetCollection<ClanMember>("membros");
-			IAsyncCursor<ClanMember> query = await Member.FindAsync(x => x.DiscordID == ctx.Member.Id);
-			List<ClanMember> responce = query.ToList();
+			IMongoCollection<Player> Player = Local.GetCollection<Player>("players");
+			IAsyncCursor<Player> query = await Player.FindAsync(x => x.DiscordID == ctx.Member.Id);
+			List<Player> responce = query.ToList();
+			
 			if(responce.Count == 0)
-			{
-				Embed
-					.WithDescription("Só membros do clã possuem acesso aos perfis.")
-					.WithAuthor("Você não é um membro do clã!")
-					.WithColor(Colors.SagiriPink);
-				await ctx.RespondAsync(embed:Embed);
-			}
+				Player.InsertOne(new Player{DiscordID = ctx.Member.Id, CurrentXP = 0, CurrentLevel = 0, _id = new ObjectId()});
 			else
 			{
-				ClanMember membro = responce[0];
-				IMongoCollection<Profile> Profile = Local.GetCollection<Profile>("profile");
-				IAsyncCursor<Profile> query2 = await Profile.FindAsync(x => x.ClanMemberID == membro._id);
-				List<Profile> responce2 = query2.ToList();
+				Player player = responce[0];
+				IAsyncCursor<ClanMember> query2 = await Member.FindAsync(x => x.PlayerID == player._id);
+			       	List<ClanMember> responce2 = query2.ToList();
+
 				if(responce2.Count == 0)
 				{
-					Profile perfil = new Profile();
-					perfil.ClanMemberID = membro._id;
-					perfil._id = new ObjectId();
-					
-					Profile.InsertOne(perfil);
 					Embed
-						.WithDescription("eu criei. Digite o comando novamente para salvar.")
-						.WithAuthor("Você não tinha um perfil, mas ...")
-						.WithColor(Colors.SagiriGray);
+						.WithDescription("Só membros do clã possuem acesso aos perfis.")
+						.WithAuthor("Você não é um membro do clã!")
+						.WithColor(Colors.SagiriPink);
 					await ctx.RespondAsync(embed:Embed);
 				}
 				else
 				{
-					Profile p = responce2[0];
-					Embed
-						.WithDescription("Digite a letra referente a sua cor:\n\n**[A]**marelo\n**[V]**erde\nA**[z]**ul\n**[L]**aranja\n**[R]**oxo")
-						.WithAuthor("Escolha a cor")
-						.WithColor(Colors.SagiriGray);
-					DiscordMessage botmsg = await ctx.RespondAsync(embed:Embed);
+					ClanMember membro = responce2[0];
+					IMongoCollection<Profile> Profile = Local.GetCollection<Profile>("profile");
+					IAsyncCursor<Profile> query3 = await Profile.FindAsync(x => x.ClanMemberID == membro._id);
+					List<Profile> responce3 = query3.ToList();
+					if(responce3.Count == 0)
+					{
+						Profile perfil = new Profile();
+						perfil.ClanMemberID = membro._id;
+						perfil._id = new ObjectId();
 					
-					DiscordMessage msg = await ctx.GetResponce(Program.Interactivity);
-
-					switch(msg.Content.ToLower())
-					{
-						case "a" :
-							p.CsColor = "A";
-							break;	
-						case "v" :
-							p.CsColor = "V";
-							break;
-						case "z" :
-							p.CsColor = "Z";
-							break;
-						case "l" :
-							p.CsColor = "L";
-							break;
-						case "r" :
-							p.CsColor = "R";
-								break;
-						default :
-							p.CsColor = "";
-							Embed
-								.WithDescription("Tenha certeza que digitou apenas a letra em destaque.")
-								.WithAuthor("Cor Inválida")
-								.WithColor(Colors.SagiriPink);
-							await botmsg.ModifyAsync(embed:Embed.Build());
-							
-							break;
-					}
-					if(!String.IsNullOrWhiteSpace(p.CsColor))
-					{
-						Profile.UpdateOne(Builders<Profile>.Filter.Eq("ClanMemberID",membro._id),Builders<Profile>.Update.Set("CsColor",p.CsColor));
+						Profile.InsertOne(perfil);
 						Embed
-							.WithDescription("")
-							.WithAuthor("Cor atualizada.")
-							.WithColor(Colors.SagiriBlue);
-						await botmsg.ModifyAsync(embed: Embed.Build());
+							.WithDescription("eu criei. Digite o comando novamente para salvar.")
+							.WithAuthor("Você não tinha um perfil, mas ...")
+							.WithColor(Colors.SagiriGray);
+						await ctx.RespondAsync(embed:Embed);
 					}
+					else
+					{
+						Profile p = responce3[0];
+						Embed
+							.WithDescription("Digite a letra referente a sua cor:\n\n**[A]**marelo\n**[V]**erde\nA**[z]**ul\n**[L]**aranja\n**[R]**oxo")
+							.WithAuthor("Escolha a cor")
+							.WithColor(Colors.SagiriGray);
+						DiscordMessage botmsg = await ctx.RespondAsync(embed:Embed);
+						
+						DiscordMessage msg = await ctx.GetResponce(Program.Interactivity);
+	
+						switch(msg.Content.ToLower())
+						{
+							case "a" :
+								p.CsColor = "A";
+								break;	
+							case "v" :
+								p.CsColor = "V";
+								break;
+							case "z" :
+								p.CsColor = "Z";
+								break;
+							case "l" :
+								p.CsColor = "L";
+								break;
+							case "r" :
+								p.CsColor = "R";
+								break;
+							default :
+								p.CsColor = "";
+								Embed
+									.WithDescription("Tenha certeza que digitou apenas a letra em destaque.")
+									.WithAuthor("Cor Inválida")
+									.WithColor(Colors.SagiriPink);
+								await botmsg.ModifyAsync(embed:Embed.Build());
+							
+								break;
+						}
+						if(!String.IsNullOrWhiteSpace(p.CsColor))
+						{
+							Profile.UpdateOne(Builders<Profile>.Filter.Eq("ClanMemberID",membro._id),Builders<Profile>.Update.Set("CsColor",p.CsColor));
+							Embed
+								.WithDescription("")
+								.WithAuthor("Cor atualizada.")
+								.WithColor(Colors.SagiriBlue);
+							await botmsg.ModifyAsync(embed: Embed.Build());
+						}
 
-				 	
+					}
 				}
 			}
